move service based gcs client file path
